backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
pandas==2.1.3
numpy==1.26.2
imbalanced-learn==0.11.0

# Data Visualization Support
matplotlib==3.8.2
seaborn==0.13.0

# OCR and Document Processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
python-docx==1.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# API and Data Handling
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# AI/LLM Integration
openai==1.3.7
anthropic==0.7.8
langchain==0.1.0
langchain-openai==0.0.2

# Statistical Analysis
scipy==1.11.4
statsmodels==0.14.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
redis==5.0.1
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# CORS and Security
python-cors==1.0.0
